Add detection code for what model of AHA154X/Z we are running on, then use that information to fix the enhancemode features of the 1542C/CF boards by turning them off.
When doing this I found that the Buslogic 545S does NOT properly
mimic the 1542 families AHA_INQUIRE command.  It only returns 1
byte of information, when the adaptec manual clearly states that 4
bytes are to be returned.  I added a printf that explains the error
when we see a 545S for now, I tried to come up with a better solution,
but it involved to much work for now.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@606 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f