From: Chris Wiener <cwiener@CRLABS.COM>
When the keyboard is probed, the LED's blink quickly and
"Keyboard reset failed" is printed on the console.  The
init routine keeps trying endlessly with the same behavior
as above.

I got the latest -current sup sources (06-Feb-94 12:00 GMT) to work using the
old syscons.c.  The following patch makes the new syscons work:


git-svn-id: file:///home/git/svn/freebsd_mirror/head@1105 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f