Work around FIFO bug in SMC UARTS.
Fix endless loop in siopoll() for an event on a tty with no tty struct.
Don't generate unwanted interrupts in the serial console driver.  These
bugs probably don't matter unless the tty struct is dynamically allocated.

Support polled mode.  To use it, leave out the irq and the vector in
the config file.  It only causes extra overhead for open polled ports.
The maximum usable speed is approximately 1000 bps for a 16450 and
15000 bps for a 16550.

Other cosmetic changes.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@3548 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f