NFS-diskless works.  Look in sys/i386/boot/netboot for some of the explanation.  More doc needed, but not hard to do, if you want to.
A big hand to Martin Renters for the netboot program !

Anybody want to compete on who can "make world" in the shortest
amount of time ?  I have 127 i486DX2/66 and 5 P60's I can use
now.  And 3 times 66 Gb file servers to support it... :->

Anyway, NFS will be standard in the GENERIC kernel now, so that
people can use the bin-tarball to set up shop.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@3794 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f