Test test test!  sigh!
Perl's scripts are still trying to execute perl out of /usr/gnu/bin/perl.

The hack Larry was using for h2ph.1 doesn't work with the new macros, so
make it a real man page.

Also, we weren't building the .ph files, add them as an afterinstall rule
in the x2p subdirectory.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@3985 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f