Add support for filesystem-specific `-o' options, and re-implement the most common cd9660 and nfs options like God intended them.  (It is now possible to say
	mount -o ro,soft,bg,intr there:/foo/bar /foo/bar

again.)  This whole getmntopt() business is an incredible botch;
it never should have been anything more than a wrapper around
getsubopt(3).  Because if the way the current hackaround is implemented,
options which take arguments (like the old `rsize' and `wsize') are still
unavailable, and must be accessed the new, broken way.

(It's unimaginable how Berkeley managed to screw up one of the few things
about NFS that Sun actually got right to begin with!)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@4065 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f