New forms library. This provides some basic functions for writing input forms. It has the following simple fields:
Text fields: Just titles, labels etc.

Input fields: An editable text field that may or may not have an
initial default value.

Labelled input field: This is an input field that has an initial
informative entry in it but it vanishes when you start editing the
field.

Toggle fields: These are fields with a pre-defined list of options
which you cycle through using the space bar.

Action fields: These are button type fields that call functions when
they are selected.

A simple demo is included in examples.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@5461 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f