Undo most of the changes in the gcc-2.6.3 "upgrade" - Don't define NO_IMPLICIT_EXTERN_C here.  It is already the default   (defined in i386/bsd.h). - Don't lose the fixed comment about FUNCTION_PROFILER. - Don't lose the define of NO_PROFILE_DATA.
Replace the unused define of COMMENT_BEGIN by the less-unused define
of ASM_COMMENT_START.  COMMENT_BEGIN was only defined in i386-specific
files and was not used in any part of gcc-2.6.3.  ASM_COMMENT_START
is defined for several targets and is used for stuff that we don't
support (dwarf).


git-svn-id: file:///home/git/svn/freebsd_mirror/head@7023 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f