Made this script a bit more fool-proof, so people like me can better use it. :-)
It now explicitly requires the specification of a directory to import
from, either as an argument to the script, or by asking the user about
it.  (Previously, it implicitly used `.', like cvs import does.)

Also implemented an option `-n', which does essentially the same like
the overall CVS option `-n': show only what would have been done,
don't do any commitment.  Note that since the modules' database is
checked out in place (and not commited back), it will erroneously be
reported as to be imported, too:

cvs import: Importing /home/ncvs/ports/foobar/foo/modules
I ports/foobar/foo/modules/CVS
N ports/foobar/foo/modules/modules

This is an unwanted side-effect, but gives the user the option to see
if the `ed' magic did the right thing when editing modules/modules.

Rod, can you please check the function ``checktag'' in the script if it
will be restritctive enough?


git-svn-id: file:///home/git/svn/freebsd_mirror/head@7396 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f