Start populating the handbook with useful cross reference targets.
All cross reference labels start with name of the file that contains
them.  A label for the top section level is simply the name of the
file (omitting the .sgml).  Other references within the file append a
colon and onother name.  For example, the label on the mailing list
section in the file eresources.sgml is eresources:mail.  This gives
each file its own cross reference namespace.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@9380 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f