stty fails with a non-obvious error message if it's run on a non-tty stream, such as a rsh or vi pipeline.
The error message is:
stty: TIOCGETD: Operation not supported

It's immediately obvious to the knowledgable hacker type, but not
exactly comforting to the user who's not native to unix.  It's
especially confusing if there's a stty command in their .cshrc and
it's showing up on rsh output.

(Fixes PR #bin/573)

Submitted by:	peter@haywire.dialix.com (Peter Wemm)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@9384 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f