This is a major rework of newvers.sh to put it back much closer to what CSRG had, plus make things like, TYPE, REVISION, and BRANCH easy to set, and derive RELEASE and VERSION from them.
Kill the JUST_TELL_ME hack, it is no longer needed.

Kill DISTNAME, I could find no reveference to it any place in the
source tree.

Now I just need to rework a few bits in release/Makefile, but want
to wait and talk to jkh about that.

Oh, and your now all running:
TYPE="FreeBSD"
REVISION="2.2"
BRANCH="CURRENT"

and the -BUILD-yymmdd is dead and gone.  The date was already in the
version[] string, no need for it to be there in 2 formats!


git-svn-id: file:///home/git/svn/freebsd_mirror/head@9515 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f