Take the ypchfn/ypchsh stuff that was removed from passwd and graft it into chpass.
Chpass can now tell when it's being asked to operate on an NIS
user and it displayes the appropriate message in the editor
template ("Changing NIS information for foo"). After the changes
have been made, chpass will promte the user for his NIS password.
If the password is correct, the changes are committed to yppasswdd.

Hopefully, this should make NIS more transparent to the end user.

Note that even the superuser needs to know a user's password before
he can change any NIS information (such is the nature of yppasswdd).
Also, changes to the password field are not permitted -- that's what
yppasswd is for. (The superuser may specify a new password, but
again, he needs to know the user's original password before he can
change it.)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@10050 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f