Call output process in slstart() whether or not there is any output.  As in pppstart(), the output process may be overloaded to handle hardware flow control and hardware output completions.
Don't wait for output to drain in slclose().  Discard output immediately.
New output is not prevented while processes are waiting for output to
drain (this is a bug), so the wait was sometimes forever.  Infinite
waits are also possible when CCTS_OFLOW is enabled and CTS is down.
Infinite waits were also caused by the above bug in slstart().

Start changing new init stuff yet again: rename unused arg `dummy'.

Reviewed by:	davidg


git-svn-id: file:///home/git/svn/freebsd_mirror/head@10659 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f