Remove private maninstall target.  It will conflict with the one in bsd.man.mk when I change the latter to use `::' instead of `:' dependencies.  (bsd.man.mk is included because NOMAN isn't defined. The maninstall target is supposed to be private to bsd.man.mk so bsd.man.mk doesn't bother testing if it is already defined.  The test for redefinition in Makefile.dev was too early to do anything.)
Change install target to a beforeinstall target (perhaps there should
be an `extrainstall' target so that Makefiles don't have to abuse
one of beforeinstall, install, realinstall or afterinstall).  Don't
bother testing for the install target already being defined.  Rewrite
the shell loop as a make loop (this reduces the time for installing
groff from 78s to 65s here).


git-svn-id: file:///home/git/svn/freebsd_mirror/head@11619 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f