Set SCSIGO generically before we determine the phase instead of doing it in each phase routine.  Saves a few instructions.
Be more careful in how we deal with SXFRCTL0.  Or in the control bits of
interest instead of using mvi.  The kernel driver will set the ULTRAEN
bit of SXFRCTL0 if we are using Ultra (20MHz) mode and we don't want to
clobber it.

In sdtr_to_rate divide by two if we are in ultra mode to get the correct
setting since its a 20MHz instead of 10MHz scale.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@11829 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f