symorder appears to have been designed to run on executable files only, as it payes no attention to the relocation table (which references the symbols).
As a result, running "symorder -c" to clean up the visibility of a LKM
".o" file (as is done in the new bsd.kmod.mk) totally screws up the
relocation table, making the LKM file unloadable. (ld: bogus
relocation record)

This is a pretty crude fix - I've changed symorder so that when
running in "cleanup" mode, it disables the reordering which was
screwing up the relocation table.  I'm sure there is a better fix, but
I didn't have the energy.  Feel free to fix this hack, probably by
renumbering the symbol indexes in the relocation table.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@11856 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f