Update the FAQ.
Is there anyone who could give me an account on a faster-than-20-bytes-per
second line to the US (not EBONE) ? I'm having too many problems connecting
to freefall to do commits... It would be better if it could be a ssh-protected
account, that would give a strong authentication.

The RENATER line (and thus EBONE) is almost dead for remote login after
10AM european time... :-(

Thanks in advance.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@12253 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f