Add -DNOCLEAN option which totally and utterly disables any form of cleaning during a make.  This may give you more rope to hang yourself if you are caught with some subtle dependency on installed binaries in your build, but if you are doing daily 'make -DNOCLEAN world' it's not too bad at all.  It could take as little as 30 minutes to do an entire sync-up of your binaries if everything's up to date, especially if you are using 'INSTALL=install -C' in /etc/make.conf (highly recomended!).
Also, add a "reinstall" target. You can do a 'make DESTDIR=/mnt reinstall'
where /mnt is the nfs root of a machine and you get the install parts of
the make world run on it.

I saw this on -hackers quite some time ago and included it in my Makefile
and have been using it on and off for a while. Alas, I cannot find the
actual message with the author's name...


git-svn-id: file:///home/git/svn/freebsd_mirror/head@12483 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f