Moved the declaration of boolean_t from <vm/vm_param.h> to <sys/types.h> (if KERNEL is defined).  This allows removing bogus dependencies on vm stuff in several places (e.g., ddb) and stops <vm_param.h> from depending on <vm_param.h>
Added declaration of boolean_t to <vm/vm.h> (if KERNEL is not
defined).  It never belonged in <vm/vm_param.h>.  Unfortunately,
it is required for some vm headers that are included by applications.

Deleted declarations of TRUE and FALSE from <vm/vm_param.h>.  They
are defined in <sys/param.h> if KERNEL is defined and we'll soon
find out if any applications depend on them being defined in a vm
header.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@12710 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f