Make pipe() return a set of bidirectional pipe fd's rather than one-way only just like on SVR4.
This has no effect on any current programs in our source, but makes
the use of SVR4 code a little easier.  There is no code or implementation
cost in the kernel.. This two-line change merely sets the modes on the ends
of the pipes to be bidirectional.  There are no other changes.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@13146 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f