Eliminate the dramatic TCP performance decrease observed for writes in the range [210:260] by sweeping the problem under the rug.  This change has the following effects:
1) A new MIB variable in the kern branch is defined to allow modification
of the socket buffer layer's ``wastage factor'' (which determines how
much unused-but-allocated space in mbufs and mbuf clusters is allowed
in a socket buffer).

2) The default value of the wastage factor is changed from 2 to 8.
The original value was chosen when MINCLSIZE was 7*MLEN (!), and is not
appropriate for an environment where MINCLSIZE is much less.

The real solution to this problem is to scrap both mbufs and sockbufs
and completely redesign the buffering mechanism used at both levels.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@13267 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f