Change calendar to report the actual date for variable day events. E.g. for Easter, and entries like "04/SunFirst" calendar will now report:
04/05*	Good Friday (2 days before easter)
04/07*	First Sunday...

instead of:

Easter-2	Good Friday...
04/SunFirst	First Sunday...

I also modified the calendar files to use the variable day format
for a lot of events so that they will be reported correctly.
E.g. U.S. daylight savings time is now listed as:

04/SunFirst	Daylight savings time...

There are still a lot of wrong dates in there for some events
that move from year to year, but I don't have a good calendar handy
right now that I can use for reference.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@15066 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f