Print some warnings if root invokes chpass and sets the shell to one of the following:
- a non-existent file
- a non-regular file
- a file without any execute bits set

The shell is still set to whatever they entered even if the above
conditions exist (hey, it is the super user doing this after all :-),
but this might give the admin. some warning that they are about to screw
themselves and give them a chance to fix it before it is too late.

Inspired by: some new FreeBSD user on USENET who set his root shell
  to a shell that doesn't exist and now can't gain access to root (don't
  worry, I sent him some mail on how to recover from this).


git-svn-id: file:///home/git/svn/freebsd_mirror/head@15181 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f