Limit the risk of `buf' overrun in ping.c when printing hostnames.
Note, this is not really a security risk, because the buffer in question
is a static variable in the data segment and not on the stack, and hence
cannot subert the flow of execution in any way.  About the worst case was
that if you pinged a long hostname, ping could coredump.

Pointed out on: bugtraq  (listserv@netspace.org)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@17320 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f