Finally implement a simple commandline history in DDB.
Emacs-style line editing has already been there (did anybody ever
notice this? :), so i `only' had to add ^P and ^N.  The approach is
fairly minimalistic, with the advantage of keeping the bloat as small
as 864 bytes of .text and 16 bytes of .bss, plus 10*120 bytes
malloc'ed history buffer at the first use.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@17495 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f