Add new option to chpass:  -e "expire" ; change the account expire time from a script as if it was done in the interactive editor.
When reassembling the gecos string, trim any excess trailing commas, they
look ugly in the passwd file. :-)

Have a simple Makefile tweak to prevent mortal users from changing their
fullname.  As ISP's we have seem some real bizzare stuff here..

When decoding the change/expire string, allow the month number as a
synonym for the name of the month.. (ie: 1 as well as Jan or January)
Note that using numbers means there's a chance that you can get bitten
if you're not used to the American DD-MM-YY order.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@17544 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f