Hooks to attempt to get 'make world' working with gcc-2.7.2.1.  It's worked for me, but no guarantees yet.
Basically, move cc earlier to lib-tools target, before libraries are
built.  libgcc is built at the same time, so it's special entry is
removed from the libraries target.  Also, add bison/gperf immediately
before cc, since they are needed to build cc and are not present on an
earler-than-2.2 system.  (libgcc is partly built with the freshly compiled
gcc-2.7.2.1, since it uses __attribute__'s that were not present until now)

This _should_ be enough to get running from older releases, perhaps as far
back as 2.0.5.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@18392 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f