Implement a -c option to ps to display the short command name instead of the full argument vector.
I've bumped into a few things that expected this switch to be present,
the most recent was the snmp package in ports.  I'm not 100% sure of the
origins of this, but Linux has it, so does the "BSD-compatable" version
of ps on our SVR4 systems (so I assume SunOS has it too).


git-svn-id: file:///home/git/svn/freebsd_mirror/head@19068 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f