Support the ".p2align" directive, which is standard in newer versions of binutils.  For all architectures and object file formats, ".p2align n" aligns to the next multiple of 2**n.  Thus for FreeBSD, it does exactly the same thing as the traditional ".align".
The old ".align" directive has different meanings in different
object formats, and even in different variants of a.out.  Sometimes
is aligns to a multiple of n, and other times it aligns to a multiple
of 2**n.  ".p2align" is preferable for use in assembly language
sources, since it makes them more portable to object formats other
than a.out.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@24355 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f