Fully implement the clause in Appendix B.4.2.2 from Posix 1003.1 that allows traditional BSD setuid/setgid behavior.
The only visible difference should be that a non-root setuid program
(eg: inn's "rnews" program) that is setuid to news, can completely
"become" uid news. (ie: setuid(geteuid())  This was allowed in
traditional 4.2/4.3BSD and is now "blessed" by Posix as a special
case of "appropriate privilige".

Also, be much more careful with the P_SUGID flag so that we can use it
for issetugid() - only set it if something changed.

Reviewed by: ache


git-svn-id: file:///home/git/svn/freebsd_mirror/head@24448 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f