Revert some changes to the selection and reselection code that were thought to fix a selection timeout problem.
If we can't find an SCB for the reconnecting target, issue a bus device
reset as the SCSI2 spec suggests.

Add a missing call to "add_scb_to_free_list" in the non paging case.  In
the non-paging case, the SCBs don't really need to be on the free list,
but putting them there clears the tag field which is something the recovery
code depends on.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@25005 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f