Submitted by:   Rich Murphey (ages ago) and Gene Stark Hopefully I've done the proper magic to merge changes between 1.17 and 1.17.2.1 into the main trunk.  Description of those changes follows:
Brought in changes sent to me in late 1995 by Rich Murphey.
I cleaned up a few things and am currently running these under
2.2-970205-GAMMA.

The changes deal with software debouncing apparently necessary on
todays faster hardware, and also some problems with the use of the -Select
line for the TW-523 sync.  This driver allows use of +PaperEnd as an
alternative.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@26292 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f