Fix problem with early revision 53c825a and 53c875 chips, which could cause a solid system lockup in the driver attach:
These chips do not abort an access to the internal SRAM, when
the driver set the software reset bit in the istat register. But
the chip will never acknowledge the requested PCI bus transfer
in the situation, causing an infinite wait and a lockout of other
bus-masters.

The problem has been reported for rev 0x11 of the 53c825a and
rev 0x01 of the 53c875.
Revisions 0x13 of the 53c825a and 0x03 of the 53c875 are known
to support SRAM accesses, even in the software reset state.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@28958 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f