Fix the broken VT_WAITACTIVE ioctl for the case where the third arg was 0.
PR:		4164
Submitted by:	Joe Traister <traister@mojozone.org>

While i was at it, also fixed a broken return value for the VT_RELDISP
ioctl, iff the third arg was legally VT_TRUE, but the destination
screen was in process mode so the actual switch had to be deferred.
This was breaking the ability to directly toggle between two X servers
running on two VTs, since the server getting the bogus error return
was running wild, and competing with the other one for the hardware.
(Sigh, this was a very long-standing bug.)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@29400 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f