Add support to the host driver for "downloaded constants".  These are immediate operands that are set during seqeuncer program download instead of at assembly time.
Remove the REJBYTE sram variable.  The host driver can just as easly
read the accumulator to get this value.

This should fix the reported problems with cards that don't have devices
attached to them failing with a stream of "Somone reset bus X" messages.

Doug Ledford determined the cause of the problem, fixes by me.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@29898 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f