Since i've just touched <sys/syslog.h>, thereby causing a large part of the system to be rebuilt anyway, this is a good time to introduce LOG_NTP.
The reasoning for a separate facility is that xntpd can sometimes
cause exaggerative log message at high prioritites which are,
depending on your environment and available clock sources, not
necessarily as important as other LOG_DAEMON messages.  However, they
used to clutter log files and system console in the existing setup.
Note that this situation could not be resolved using the !xntpd option
(think about it).

xntpd(8) is supposed to automatically pick up the change, it had
already all necessary #ifdef's.

The chosen value does, as far as my inquiries yielded, not clash with
any other operating system.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@30183 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f