Fix two bugs which caused various RPC programs (mountd, nfsd, ...) to fail under certain circumstances.
1. In one spot, the ifr_flags member was being examined in the
wrong structure, thus it contained garbage.  On a machine in which
only the loopback interface was up, this caused everything that
wanted to talk to the portmapper to fail -- a particular problem
with laptops, where the pccard ethernet interface is likely to come
up long after the attempt to start mountd, nfsd, amd, etc.

2. Compounding the above problem, get_myaddress() returned a
successful status even though it failed to find an address that it
considered good enough.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@30504 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f