Fix the CCP Type field value for DEFLATE. (I *really* meant to do this  *before* committing the deflate changes in the first place - oops).
Pppd is horribly broken in this respect - refer to the
ppp man page for details.  Ppp *WON'T* negotiate deflate
with pppd by default - you must ``enable'' and ``accept''
``pppd-deflate'' in your config.

While I'm in there, update the cftypes in ccp.c so that
we recognise some more protocols (we don't actually do
anything with them - just send a REJ).


git-svn-id: file:///home/git/svn/freebsd_mirror/head@31518 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f