Add new (very useful) tunable for pageout daemon.  The flag changes the maximum pageout rate:
sysctl -w vm.vm_maxlaunder=n

 1 < n < inf.

If paging heavily on large systems, it is likely that a performance
improvement can be achieved by increasing the parameter.  On a large
system, the parm is 32, but numbers as large as 128 can make a big
difference.  If paging is expensive, you might try decreasing the
number to 1-8.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@31550 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f