Reverse my previous change and use htons() on an int instead of htonl() !
This results in the int a,b,c,d changing to b,a,c,d,
but as it's subsequently coerced to a u_short, the
ultimate answer is correct.

If this isn't fixed properly soon (by the author) I'll
have a look at it again.

Noted by:	eivind & ari@suutari.iki.fi


git-svn-id: file:///home/git/svn/freebsd_mirror/head@31576 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f