If perl exists in OBJDIR it has just been built; if it exists in CURDIR it has been built without an obj directory; however if it is in neither of those places, we expect it to be in DESTDIR.
Yes Bruce, I know this is broken because the host is not supposed to be
the same as the target, but we need to get the hosted build working
properly first before even attempting a cross compiled operating
system build. That will need to concept of TOOLSDIR or something that
can be mapped to DESTDIR in the case of a hosted build and set to the
installed tools in a cross compiled build. Later, later, later!


git-svn-id: file:///home/git/svn/freebsd_mirror/head@35985 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f