Since I got no objections to this patch, and no one has offered any alternative, I present .. ta! da! .. the __error() hack.
This patch to the a.out dynamic loader provides old a.out binaries
with __error() if they are linked with an older libc that lacks it,
but are also linked against a library that needs it.

There is a smaller, tricker hack that takes advantage of the fact
that ld.so has __error() too, courtesy of the new libc, but this
hack is the straightforward version.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@37092 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f