Cast to `char *' instead of to u_long to help add byte offsets to pointers.  Neither is portable, but "correct" casts to integral types are much uglier - they lead to expressions like
        ptr = (struct struct_with_too_long_a_name *)(void *)(uintptr_t)
	    ((uintptr_t)(void *)ptr + offset);

Here the cast to the struct pointer is to match the surrounding
style of this file (and not depend on C's feature of properly
converting `void *' on assignment or cast), the `void *' casts are
because uintptr_t is only specified to work on `void *' pointers
(I missed this in about 100 lines of previous changes from [u]long
to [u]intptr_t), the outer cast to a uintptr_t is in case the
addition promoted the type, and the inner cast to a uintptr_t
corresponds to the one cast to an integer in the original code.

Don't depend on gcc's feature of casting lvalues.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@38346 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f