Honour the default CFLAGS.  This fixes a spew of warnings for compiling unused static inlines in headers without -O.
Fixed missing libraries in DPADD.

Use .PATH instead of a symlink farm.

Removed bogus dependency of util.c on statetext.h (only util.o depends
on it).  Removed rule for building statetext.h.  The dependency was used
to get statetext.h built, but statetext.h is a non-generated source file
so it doesn't need building.  This fixes an annoying message for the null
build and use of `make' instead of ${MAKE} to do the null build.

Fixed some style bugs.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@40489 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f