GRRRR! Apparently, the promiscuous mode chip bug which I thought was isolated to revision 33 PNIC chips is also present in revision 32 chips. Cards with rev. 32 chips include the LinkSys LNE100TX and the Matrox FastNIC 10/100. This accounts for all the cards that I have to test with.
(I was never able to personally trip the bug on this chip rev, but today
one of the guys in the lab did it with the software they're working on
for their cellular IP project, which uses BPF and promiscuous mode
extensively.)

This commit enables the promiscuous mode software workaround code for
both revison 32 and revision 33 chips. It's possible all of the PNIC
chips suffer from this bug, but these are the only two revs where I
know for a fact it exists.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@42296 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f