Rewrite of crypt library to be more modular, and addition of the Secure Hashing Algorithm - 1 (SHA-1), along with the further refinement of what $x$salt$hash means.  With this new crypt the following are all acceptable:
    $1$
    $MD5$
    $SHA1$

Note: $2$ is used by OpenBSD's Blowfish, which I considered adding
as $BF$, but there is no actual need for it with SHA-1.  However,
somebody wishing to add OpenBSD password support could easilly add
it in now.

There is also a malloc_crypt() available in the library now, which
behaves exactly the same as crypt(), but it uses a malloced buffer
instead of a static buffer.  However, this is not standard so will
likely not be used much (at all).

Also, for those interested I did a brief speed test Pentium 166/MMX,
which shows the DES crypt to do approximately 2640 crypts a CPU second,
MD5 to do about 62 crypts a CPU second and SHA1 to do about 18 crypts
a CPU second.

Reviewed by:	Mark Murray


git-svn-id: file:///home/git/svn/freebsd_mirror/head@42981 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f