Add a new sysctl node kern.shutdown, off which shutdown-related things can be hung.
Add a tunable delay at the beginning of the SHUTDOWN_FINAL at_shutdown
queue, allowing time to settle before we launch into the list of things
that are expected to turn the system off.

Fix a bug in at_shutdown_pri() where the second insertion always put
the item in second position in the queue.

Reviewed by:	"D. Rock" <rock@cs.uni-sb.de>


git-svn-id: file:///home/git/svn/freebsd_mirror/head@43436 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f