Disable tagged queueing for the IBM DCAS drives.  These drives have poor write performance when tagged queueing is enabled.
Although the PR was submitted for the 4 gig version of this drive, the
assumption is that the 2 gig version has the same problem.  Therefore
tagged queueing is disabled for both.

Also, update the comment for the Western Digital Enterprise drives to note
that the best performance for those drives is achieved when tagged queueing
is disabled and write caching is enabled.

PR:		kern/10398
Submitted by:	Hideaki Okada <hokada@isl.melco.co.jp>


git-svn-id: file:///home/git/svn/freebsd_mirror/head@44731 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f