Fixed a couple of missing drive selects in wdreset().  This mainly affects cases where there is a slave but no master.  These bugs were usually recovered from provided ATAPI was configured, but only after lengthy delays.  Configuring ATAPI still fixes some bugs for non-atapi drives.
Don't wait for seek complete in wdreset().  If necessary for pre-ata
drives, it can be waited for later (we got it wrong by only looking
at it for drive 0 anyway).  It is set as part of a historical
signature for ata drives but doesn't say anything useful about the
reset state.  It is cleared as part of a non-historical signature
for atapi drives so that drivers which don't understand atapi drives
seem to see no drive.  Waiting for it caused lengthy delays and
broke the status returned by wdreset() in cases where the master
was not an ata drive.  Then the whole wdprobe() failed in some
cases where the recovery code didn't work.

Don't wait for drive ready in wdreset().  The considerations are
the same as for seek complete, except drive ready does say something
useful about the reset state of ata drives, and waiting for it
later is required anyway for such drives.

Lengthy delays can now be avoided by not configuring nonexistent
(ata) drives.  Unfortunately, this breaks detection of atapi drives
in some configurations.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@44971 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f