I'm sick of the automatic rename of /kernel to /kernel.old while doing development that leads to lots of crashes during boot.
I have made a 'reinstall' target (like in ports, and reinstall.debug)
This is most useful if you want to keep /kernel.old as a known bootable
kernel.  If you test a new kernel and have to reboot for a fix, a
'make reinstall' will install the new kernel over the top of the old
non-viable one, leaving the old one untouched.  This is mainly meant
for development, not general users.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@46726 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f