Zap totally the npx0 memory size override.  It only worked if statically specified in the kernel config file - but setting options MAXMEM works exactly the same.  Userconfig overrides of this have not worked for ages.
Also, change the getenv for the loader override to hw.physmem based on a
prior suggestion from Mike Smith.  I think he still wants to change this
some, but this shouldn't get in his way.  This is a forced setting of
the memory size, not a "cap".  We probably should have a plain 'maxmem'
variable as well which does do a cap, without loosing the bios memory
configuration data.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@48445 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f