Revert some changes I had made to try and cut down on the number of TX EOF interrupts that were scheduled. Testing shows it didn't really do very much and it makes the code a little more complicated (which is never a good thing).
Also fix the rambuffer offset initialization for the 512K/64K SRAM case
(512K total using 64K chips). It should be 0. The only case with a
non-standard rambuffer offset address is 1024K/64K according to the
SysKonnect manual. (My card has the 1024/64 configuration and I don't know
which card uses the 512/64 configuration, if any, so I'm not sure that
this was really a problem for anyone.)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@48812 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f