Make jdk-1.1.8 work again.  It turns out that some code inside libjava peeks into the dynamic linker's private Obj_Entry structures. My recent changes introduced some new members near the front of the structures, causing libjava to get the wrong fields.  This commit moves the new members toward the end of the structure so that the layout of the portion that is relevant to JDK remains the same as before.
I will work with the JDK porting team to see if we can come up with
a less fragile way for them to do what they need to do.  I understand
the current approach was necessary in order to work around some
limitations of the dynamic linker.  Maybe it's not necessary any
more.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@50977 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f