Don't clear the SA_FLAG_TAPE_LOCKED prior to calling saprevent when we're doing an OFFLINE ioctl- saprevent won't unlock the door, which then causes the unload to fail on some units.
If we've already unloaded the tape, don't try and rewind and
unload in saclose. This is a slightly riskier change because
we're now going to depend on SA_FLAG_TAPE_MOUNTED to say whether
we've really unloaded the tape. This involved changing the
setting in sadone for tape errors to SA_FLAG_TAPE_FROZEN (which
is more accurate anyway-if you get an EIO you've probably lost
tape position anyway) where it used to just clear the mounted
flag.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@53522 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f