Add one of the two root words adapted from scandinavian since 1066:         ombudsman and its mate         ombudsmanship since I've had ispell (which uses /usr/share/dict/words) flag it too many times.
Thanks to phk for telling me the other word is kokkenmoding, an
archaeological term for a pile of oystershells and other debiris from
a stone-age household.  I didn't add it to the list, however, since
none of the online dictionaries that I looked at had it in them.
Gotta find that oed lookup link :-)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@53662 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f