Better error handeling:
On UDMA CRC errors retry operation as it might be a fluke, if not fall
back to PIO mode on the failing drive. If you get alot of these your
cabeling is most likely not good enough.

On HARD error using DMA, retry once using PIO, if it succeds using PIO
fall back to PIO mode on the failing drive.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@53882 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f