Update mdoc(7) and mdoc.samples(7) to list the preferred order of manual page sections.  Make the two man pages consistent with each other in the headers they list and they order they list them in.
Note:  this is the preferred ordering.  All new man pages/additions
to man pages should try and follow this.  Existing man pages
should be left alone, unless you are making major changes  in
the man page and re-ordering of the sections is only a
minor part of the change.

PR:		doc/15352, doc/15353


git-svn-id: file:///home/git/svn/freebsd_mirror/head@54337 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f