This should have been committed with related changes to .c files.
Changed the type used to represent the user stack pointer from `long *'
to `register_t *'.  This fixes bugs like misplacement of argc and argv
on the user stack on i386's with 64-bit longs.  We still use longs to
represent "words" like argc and argv, and assume that they are on the
stack (and that there is stack).  The suword() and fuword() families
should also use register_t.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@55142 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f