Add the vendor and device IDs for a whole bunch of additional USB ethernet adapters that are supported by the aue and kue drivers. There are actually a couple more out there from Accton, Asante and EXP Computer, however I was not able to find any Windows device drivers for these on their servers, and hence could not harvest their vendor/device ID info. If somebody has one of these things and can look in the .inf file that comes with the Windows driver, I'd appreciate knowing what it says for 'VID' and 'PID.'
Additional adapters include: the D-Link DSB-650 and DSB-650TX, the
SMC 2102USB, 2104USB and 2202USB, the ATen UC10T, and the Netgear EA101.
These are all mentioned in the man pages, relnotes and LINT.

Also correct the date in the kue(4) man page. I wrote this thing
on Jan, 4 2000, not 1999.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@55579 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f