Compile *_genassym.c with ordinary ${CFLAGS}.  The (small) needs for -U_KERNEL became negative when all all the genassym.c's were converted to be cross-built.  Related cleanups: PARAM went away, but was still used here; KERNEL was renamed to _KERNEL, but was still KERNEL here; the deprecated macros $@ and $< were still used here.
Use "genassym ... > ${.TARGET}", not "genassym -o $@ ...", so that
genassym(1) doesn't need to support -o.

Removed half-baked hard-coded dependencies of *_genassym.o on headers.
These objects should be added to the list of objects in the depend
rule to get full dependencies.  This doesn't happen automatically
because they are not linked into the kernel.  Half baked dependencies
don't really help.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@55642 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f