Remove garbage collected tags from their usage in the ccb_getdev structure. Remove usage of the (now gone) pd_type tag of same.
Add an extra probing state such that if we successfully run an
initial inquiry (36 bytes), rerun another one with a longer data
size as informed by the 'additional length' field in the first
returned inquiry data (making sure not to get bigger than the
actual scsi_inquiry_data structure- which has also been modified-
see separate checkin of scsi_all.h). This allows devices such
as SAF-TE devices (which have identifying marks in offsets 48-53
in inquiry data) to be successfully found without special case
inquiry commands. There are also a lot of other things such as
version codes that are coming in in the SPC2 specification that it
would be useful to get our hands on.

Reviewed by:	gibbs@freebsd.org, ken@freebsd.org


git-svn-id: file:///home/git/svn/freebsd_mirror/head@56146 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f