Don't include <machine/ipl.h> in <sys/systm.h> in the i386 case.  This fixes some namespace pollution in general and breakage of modules that aren't in the sys tree in particular (<machine/ipl.h> includes further headers that aren't installed under /usr/include).
Reimplemented SPLASSERT() so that it is more machine independent and
less bloated and doesn't require the <machine/ipl.h> include spam.
In particular, don't assume that `cpl' can be printed using %08x
format.  The alpha arch doesn't even have `cpl'.  SPLASSERT() was
harmless on alphas because it isn't actually used.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@57294 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f