Sync with an early version of the -current code. Mainly a tidy up of some function names, dropping the ISA bits that were never used and removing __P
We've renamed the _pccard_ stuff as nothing else probes or attaches.

In the next commit we will move some routines around and split out some of the
structure definitions to another file.

I've dropped the SYSCTL as I never got it working. APM is gone too as it
is broken.

Note, that the attr_read/write will be in both this and NEWCARD/NEWBUS so
I've moved the prototypes to the right place.
Add a few casts.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@60158 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f