Fix the real problem that broke the Alpha loader this last week.  It was not the fault of the module code, nor FICL.  The malloc code requires sbrk() to return addresses that were at least 16 byte aligned.  If the Alpha loader happened to be 8 byte but not 16 byte aligned in length, then you would get a zfree() panic at startup.
Incidently, this affected the i386 loader as well, and explains why
the static heap changed things and why jlemon had trouble when the bss
was not ending at a multiple of 8 bytes.

My fix is to 16 byte align it on all arches, even though the x86 version
only required 8 byte alignment (struct MemNode is smaller there).  We could
page align it if we wanted to be paranoid, but it isn't presently necessary.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@60481 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f