Add support for the National Semiconductor DP83815 fast ethernet controller chip. This chip is currently being used on the NetGear FA312-TX adapter, which I guess is a replacement for the FA310-TX (PNIC-based).
I added support for this chip by modifying the sis driver since
the SiS 900 and the NS DP83815 have almost the same programming
interface (the RX filter programming and PHY access methods are
different, but the general configuration, DMA scheme and register
layout are identical).

I would have had this done a lot sooner, but getting the damn MAC
address out of the EEPROM proved to be more complicated than expected.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@62672 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f