Make a separate build directory in which to put all the generated objects, to simplify working with PicoBSD.
Add the ability to put make instructions in crunch.inc to pass to the
build process.

Now explicitly make the objects in our own object tree, since we want to
build the objects with our own defines, and allow this to occur in a
common object tree for all PicoBSD builds, if required.  This is
controlled by the COMM_OBJ variable, for those who don't want this -
setting it to /usr/obj again will just pick up the objects from your
last make buildworld, as before.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@63425 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f