(null commit)
Reflect that the update to routed-2.22 addresses a problem discussed
in PR bin/16941.  The updated release also incorporates our local
buffer overflow fix (implemented differently) and puts us in a
position where we are completely synchronized with the master
distribution, apart from the stupid spammage of the #ident
definitions, which are supposed to indicate the release version,
not individual source file revisions.

This distribution was not moved into contrib because of objections
from wollman.  He feels that the software is original Berkeley
code that shouldn't be imported into contrib just because it
has an external maintainer.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@64526 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f