Maintain input and output throughput averages and choose the highest of the two when calculating the MP throughput average for the ``set autoload'' implementation.
This makes more sense as all links I know of are full-duplex.  This
also means that people may need to adjust their autoload settings
as 100% bandwidth is now the theoretical maximum rather than 200%
(but of course, halfing the current settings is probably not the
correct answer either!).

This involves a ppp version bump as we need to pass an extra
throughput array through the MP local domain socket.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@64670 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f