Add code to devname(3) so it can find the names of devices which were not present when dev_mkdb(8) was run.
First the dev_mkdb(8) database is searched, this caters for non-DEVFS
cases where people have renamed a device.

If that fails we ask the kernel using sysctl kern.devname if the device
driver has put a name in the dev_t.  This covers DEVFS cloned devices.

If that also fails we format a string which isn't entirely useless.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@65632 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f