Fixed the `mysterious blank lines' problem.  This was due to unconditional use of the .ne requests.  groff(1) apparently inserts some virtual trap at the end of each page, and this was causing the .ne request to output extra lines.  Solution is to only use them in troff mode.
Also fixed a related bug in .It macros, where it would insert a page break
in the middle of a sentence.

Removed the end of page trap in nroff mode, as this causes .sp requests to
sometimes eat the space due to the nearby trap.

Removed blank line that appeared as the first line on each nroff-formatted
manual page.

Obtained from:	NetBSD


git-svn-id: file:///home/git/svn/freebsd_mirror/head@66171 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f