For bus_setup_intr, cookiep is a pointer to storage for the parent bus to use.  We need to set it here.
This fixes the problem where a probe routine establishes and
disestablishes the interrupt and then we get a panic in the probe
routine.

However, we pass the pointer to the interrupt hanlder count to the
parent bus, which writes its own cookie there, so there may be some
problems with that which isn't apparent at the moment.

Commit made from: laptop running NEWCARD with sn driver (which works,
but gets the wrong ethernet address).


git-svn-id: file:///home/git/svn/freebsd_mirror/head@70761 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f