Add was_fabric_dev/fabric_dev tags to our local FC database structure (so we can see rapidly whether something was a fabric device but is now gone).
Add a tag which says what role this adapter should take. It can take
on the value of None, Target, Initiator or Both. None is useful for
warm failover purposes. Remove the ISP_CFG_NOINIT silliness since
a role of "None" does this.

Add a isp_lastmbxcmd tag to store the opcode for the last mailbox
command used.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@71079 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f