Fix the problems I (and others, undoubtedly) have been having for a while with threaded software in -CURRENT acting very "weird".  It has seemed, for example, in Mozilla that threads attempting to do host lookups have been locking up.  That's exactly the case.
There was a race condition in the implementation of the initialization
of the mutex used to protect FILE operations, first of all: multiple
instances of FLOCKFILE() in libc could occur on the same FILE at
the same time and cause strange behavior by overwriting eachothers'
creation of the mutex and the rest of the file lock.

Secondly, it's not appropriate to test the "validity" of the file
descriptor referenced by the FILE; if the code is calling FLOCKFILE()
or FUNLOCKFILE(), it wants the FILE to be locked or unlocked, not
to be locked or unlocked on the condition that _file is >= 0.  This
also could quite easily cause leaks by failing to perform the lock or
unlock operation when it actually is needed.

Mozilla now works again on -CURRENT when linked to libc_r.so.5 and
libc.so.5.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@72937 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f