Fix some very broken code in __nc_error() that implements per-thread `nc_error' variables. Move the nc_lock mutex from mt_misc.c to a static variable within this function, since it is only used here.
Add a new getnetconfigent() error code `NC_NOTFOUND' to report the
case where the specified netid was not found. Set nc_error in all
error cases in getnetconfigent() so that the error messages returned
by nc_(s)perror are always meaningful.

Add a terminating \n to the output of nc_perror() to match both
our manpage and other implementations of this function.

Reviewed by:	deischen, alfred, Martin Blapp <mb@imp.ch>


git-svn-id: file:///home/git/svn/freebsd_mirror/head@75146 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f