Reclaim directory vnodes held in namecache if few free vnodes are available.
Only directory vnodes holding no child directory vnodes held in
v_cache_src are recycled, so that directory vnodes near the root of
the filesystem hierarchy remain in namecache and directory vnodes are
not reclaimed in cascade.

The period of vnode reclaiming attempt and the number of vnodes
attempted to reclaim can be tuned via sysctl(2).

Suggested by:	tegge
Approved by:	phk


git-svn-id: file:///home/git/svn/freebsd_mirror/head@75654 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f