Add a new field 'md_kernnest' to the alpha machine dependent process structure.  This field keeps track of how many levels deep we are nested into the kernel.  The nesting level is bumped at the start of a trap, interrupt, syscall, or exception and is decremented on return.  This is used to detect the case when the kernel is returning back to a kernel context in exception_return().  If we are returning to the kernel we need to update the globaldata pointer register saved in the stack frame in case we have switched CPU's between taking the initial interrupt that saved the frame and returning.  If we don't do this fixup it is possible for a CPU to use the wrong per-cpu data.  On UP systems this is not a problem, so the code is conditional on SMP.
A count was used instead of simply checking the process status register in
the frame during exception_return() since there are critical sections at
the very start and end of a trap, exception, or interrupt from userland in
which we could trash the t7 register being used in userland.  The counter
is incremented after adn before these critical sections respectively so
that we will not overwrite the saved t7 register if we are interrupted
during one of these critical sections.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@75925 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f