Handle situations where we've already got a P2P interface address of a/x -> b and then negotiate a/x -> c by simply expecting SIOCAIFADDR to do the change.
This was broken by the last commit that optimised out the deletion and
re-addition of the same a/x -> b combination, and forgot to compare
the old/new destination addresses.

Conveniently enough, this problem didn't effect setups where the
default route goes via the ppp link, and most other setups don't
care what the the destination address is actually set to.  It broke
test environments where ppp connects to the local machine rather
badly though....


git-svn-id: file:///home/git/svn/freebsd_mirror/head@76214 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f