Simplify to bitmaskcmp() to use the obvious approach instead of comparing bit by bit.
Make the logic in in6_fillscopeid() match that in our ifconfig(8):
only set the scope ID if there is one in the address and none in
sin6_scope_id.

Correct a comment in network_init() that didn't make sense; it was
probably never updated after it was pasted from similar code in
addrmerge().


git-svn-id: file:///home/git/svn/freebsd_mirror/head@79720 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f