Rip out the old __stdin/out/err stuff. It was completely 100% useless. :-( It was foiled because of dynamic copy relocations that caused compile-time space to be reserved in .bss and at run time a blob of data was copied to that space and everything used the .bss version..  The problem is that the space is reserved at compile time, not runtime... So we *still* could not change the size of FILE.  Sigh.  :-(
Replace it with something that does actually work and really does let us
make 'FILE' extendable.  It also happens to be the same as Linux does in
glibc, but has the slight cost of a pointer.  Note that this is the
same cost that 'fp = fopen(), fprintf(fp, ...); fclose(fp);' has.
Fortunately, actual references to stdin/out/err are not all that common
since we have implicit stdin/out/err-using versions of functions
(printf() vs. fprintf()).


git-svn-id: file:///home/git/svn/freebsd_mirror/head@81600 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f