Implement a better compatibility mode with mount_mfs.  It is the default if the executable is named (called as) "mount_*", or can be enabled with the -C option.  This allows users to leave their old fstab entires unchanged (modulo symlink'ing mdmfs to mount(md|mfs)) and have things behave the way they should (by emulating mount_mfs silliness), while still allowing mdmfs to be used as a generic make-an-md-and-mount-it type thing.
Right now, the only effects of this option is to set the mount-point
mode to 01777 as if "-p 1777" was given, and to complain about getting
command-line options that mount_mfs didn't take (e.g., -X, -L, et al).
The latter is mostly to try to catch operator errors.

Also implement -U, which turns on soft-updates.  It's redundant (since
softdep is the default), but implement it anyway for compatibility.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@81742 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f