Back out the previous fix to deal with kernels that don't support IPv6, and implement a far more subtle and correct fix.
The reason behind the infinite loop was that ppp was trying to make up
initial IPv6 numbers and wasn't giving up when it failed unexpectedly to
assign the addresses it just fabricated to it's interface (thinking that
the reason was because another interface was using the same address).
It now attempts this up to 100 times before just failing and trying to
muddle along (in reality, this should never happen more than a couple
of times unless our random number generator doesn't work).

Also, when IPv6 is not available, don't even try to assign the IPv6
interface address in the first place...


git-svn-id: file:///home/git/svn/freebsd_mirror/head@81897 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f