Handle "identifier strings" right.  Each ISA PnP card must have a mandatory "card" identifier string.  A logical devices on the ISA PnP card may optionally have a "device" identifier string. Do not confuse them.
The "card" identifier string is assigned to a logical device as the
default description string when the device is found. (If the "card"
identifier string has not been found, use the EISA PnP ID string.
Strictly speaking, this is an error.)  We will override it when a
"device" identifier string is found later.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@83504 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f