Add a fix for the VIA82C686B data corruption bug. This fixed the problem on the 3 platforms I've been able to test on.
I'm still of the oppinion that the BIOS should take care of this,
however some board makers only apply this when they spot a
SBLive! soundcard, but the problem exists even without a SBLive!.

This fix should probably go somewhere else, but for now I'll
keep it here since we havn't got a central place to put
such things.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@83932 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f