Do not ``build'' (ie remove, and put in stub libraries) for libresolv and/or libgnumalloc on anything but i386.  The other platforms post-date this mistake.
Do not build libc_r for ia64.  There are some fundamental issues that
need to be resolved (ie: it cannot use setjmp/longjmp for thread
switching, which isn't likely to be fixed soon.  libc_r has to be
reimplemented using something like makecontext()/swapcontext() etc
in order to work in ia64.)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@85951 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f