* Make sure we increment pm_stats.resident_count in pmap_enter_quick * Re-organise RID allocation so that we don't accidentally give a RID   to two different processes. Also randomise the order to try to reduce   collisions in VHPT and TLB. Don't allocate RIDs for regions which are   unused. * Allocate space for VHPT based on the size of physical memory. More   tuning is needed here. * Add sysctl instrumentation for VHPT - see sysctl vm.stats.vhpt * Fix a bug in pmap_prefault() which prevented it from actually adding   pages to the pmap. * Remove ancient dead debugging code. * Add DDB commands for examining translation registers and region   registers.
The first change fixes the 'free/cache page %p was dirty' panic which I
have been seeing when the system is put under moderate load. It also
fixes the negative RSS values in ps which have been confusing me for a
while.

With this set of changes the ia64 port is reliable enough to build its
own kernels, even with a 20-way parallel build. Next stop buildworld.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@86213 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f