Use osigreturn(2) instead of sigreturn(2) plus broken magic for returning from old signal handlers.  This is simpler and faster, and fixes (new) sigreturn(2) when %eip in the new signal context happens to match the magic value (0x1d516).  0x1d516 is below the default ELF text section, so this probably never broken anything in practice.
locore.s:
In addition, don't build the signal trampoline for old signal handlers
when it is not used.

alpha:
Not fixed, but seems to be even less broken in practice due to more
advanced magic. A false match occurs for register #32 in mc_regs[].
Since there is no hardware register #32, a false match is only possible
for direct calls to sigreturn(2) that happen to have the magic number
in the spare mc_regs[32] field.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@90132 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f