Reserved one of the spare fields in struct gmon to record the history counter type, as threatened in rev.1.8 (the density doesn't need to be recorded since it can be derived from other fields).  This doesn't affect binary compatibility, but new utilities won't be able to depend on the contents of this field because libc/gmon/gmon.c was broken -- it wrote garbage to the spare fields.
Added a history counter type field to struct gmonparam.  This breaks
binary compatibility a little, since kgmon wanted to read the whole
struct.  Fixed kgmon to only depend on reading the critical earlier
parts of the struct.  This should also fix 6+ year old breakage of
binary compatibility when the profrate field was added.

Only initialize the new field in struct gmon for now, so that the
compatibility code for this (in kgmon) gets tested.  The compatibility
code has to guesstimate the value.  The new field in struct gmonparam
is for the kernel to initialize so that kgmon doesn't have to guess.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@91009 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f