Fixed some misspellings of 2 as sizeof(UNIT) so that they won't break things when sizeof(UNIT) becomes a runtime parameter.  The relevant 2 is the one in profil(2)'s scaling of pc's to bucket numbers:
	   bucket = (pc - offset) / 2 * profil_scale / 65536

gprof(1) must duplicate this scaling, bug for bug compatibly, so it
must first do an integer division by 2 although this mainly makes
scales larger than 65536 useless.  sizeof(UNIT) was already wrong in
gprof4, but there were no problems because the fake profil scale is a
multiple of 2.

There are also some rounding bugs in the scaling, but these are only
problems if profil(2) is used directly to create unusual (and not
useful) scales.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@91735 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f