Embed boot images built as part of buildworld rather than the installed ones under /boot (which we may not even have in the case of a cross build).
This introduced chicken and egg problem - we need boot images
early in the "depend" stage but they have not yet been built.
Work around this by excluding the generated makeboot.c source
from the "depend" list; it's okay because we hardcode all its
dependencies explicitly.  We actually lose the dependency bit
on <sys/types.h> but it's probably okay too as the only thing
we use is the u_char datatype and this is unlikely to change.
After all, it's normal for sloppy cleaning to cause problems.

beast.FreeBSD.org running 5.0-CURRENT alpha has been able to
cross build i386 world with this patch.

Prodded by:	gallatin


git-svn-id: file:///home/git/svn/freebsd_mirror/head@92323 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f