Teach GEOM about Sun disklabel formats.
The detection code in this method is written so that it should work on
all architectures which means that you can plug a Sun disk into a i386
now and access the partitions.

We still need an endian-agnostic ufs/ffs before this is really
interresting, but the main focus was to get sparc64 onto the GEOM
trail.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@92372 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f