Hide the ACPI counter probing printf behind bootverbose.
The conclusion is that this method really can tell the perfect from the
less than perfect ACPI counters.

It is in fact probably a bit more discriminative than that, but we
will rather condemn some otherwise perfect counters to the slightly
slower "-safe" version, than certify a counter as perfect which
will let us down later.

Many thanks to all the people who sent email reports!


git-svn-id: file:///home/git/svn/freebsd_mirror/head@93093 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f