I've fixed the X11 port, so I can remove the (ioctl) hack.
This hack turns out to be a bad idea.  Many auto configuration scripts
will #include a bunch of system headers, run it through cpp and then
grep/sed to see if there was really a prototype.  The regular
expressions that are used to test this didn't accept all possible
legal ANSI-C prototypes.  In particular, they didn't accept the ()
around ioctl.  Several different programs do this, so eliminating the
parans will keep FreeBSD maximally compatible with other systems.

Appologies to: obrien (he was right in the end)

This potentally breaks some ports (and fixes vim and others that
haven't been fixed to deal with the parens).


git-svn-id: file:///home/git/svn/freebsd_mirror/head@94345 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f