High-level changes (user visible): o  Implement -c (clear) to clear previously kept headers (note that    dumps not cleared will remain until -c is used), o  Implement -f (force) to allow re-saving a previously saved dump, o  Implement -k (keep) and make clearing the dump header the default, o  Implement -v (verbose) and make most output conditional upon it, o  Emit minimal output for the non-verbose case with the assumption    that savecore is run mostly from within /etc/rc, o  Update usage message to reflect what is and what's not, o  mark -d as obsolete.
Low-level changes:
o  Rename devname to device, for devname mirrors a global declaration
   and GCC 3.x warns about it,
o  Open the dump device R/W for clear and !keep to work,
o  Reorder the locals of DoFile according to style(9),
o  Remove newlines from strings passed to warn* and err*,
o  Use stat(2) to check if a dump has been saved before,
o  Truncate existing core and info files to support force,
o  First check for the magic and the version before we complain about
   parity errors. This prevents emitting parity error messages when
   there's no dump,
o  Keep track of the number of headers found and the number of headers
   saved to support the minimal output,
o  Close files we opened in DoFile. Not critical, but cleaner.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@94580 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f