If the receiver runs out of space for an received frame in the internal FIFO or the in-RAM descriptors it will switch to RX_IDLE from where it is not restarted.
We used to deal with RX_IDLE by doing a total reinit but this lost
our link and caused a potential 30sec autonegotiation against
switches.  This was changed to a less heavyhanded approach, but this
failed to restart the receiver it it were in the RX_IDLE state.

This change adds the RX_IDLE and the RX_FIFO_OFLOW conditions as
triggers for interrupts and receive side processing, and restarts
the receiver when it is RX_IDLE.

Remove the #ifdef notyet'ed nge_rxeoc() function.

Sponsored by:	Cybercity Internet, Denmark.
MFC after:	7 days


git-svn-id: file:///home/git/svn/freebsd_mirror/head@94612 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f