Finish adding support code for sysctl kern.mprof.  This dumps some malloc information related to bucket size effeciency.  Three things are printed on each row:
Size is the size the user actually asked for rounded to 16 bytes.
Requests is the number of times this size was asked for.
Real Size is the size we actually handed out.

At the end the total memory used and total waste is displayed.  Currently my
system displays about 33% wasted memory.

The intent of this code is to gather statistics for tuning the malloc bucket
sizes.  It is not intended to be run with INVARIANTS and it is not entirely
mp safe.  It can be enabled via 'options MALLOC_PROFILE' which was commited
earlier.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@94730 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f