- Axe -mpreferred-stack-boundary=2 as -Os turns this on by default. - Axe -fdata-sections as turning it on or off makes no difference.  If   it did make a difference it would serve to bloat boot2 even further with   extra padding. - Axe -fforce-addr.  This gets us 32 bytes so we are down to only being   64-bytes over.
We still can't compile this with gcc 3.1.  The problem seems to be that
the -fno-align-foo options don't actually work.  Comparing the new and
old output it turns out that gcc is 4-byte padding all the functions and
labels and what not despite the passed in arguments thus adding the
unfortunate bloat to boot2.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@96327 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f