Remove original license disclaimer. Add Caldera license.
Approved by:    David Taylor <davidt@caldera.com>

Make buildable under FreeBSD.  This was relatively complicated: the
original text used the msU macros, which are available in a number of
different kinds.  This version uses a number of mm-like macros,
including AL and BL, which just aren't available in ms, and the msU
macros I've found (even the ones in 4.4BSD) don't have them either.  I
replaced them with ms constructs, which makes it format better than
the O'Reilly document, but I wasn't able to get the table of contents
(ms doesn't have that facility).

Add a prominent comment that this is not a reference for any modern
version of C.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@96931 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f