Almost complete rewrite of the lpc commands 'abort', 'enable', 'disable', 'restart', 'start', 'stop' and 'up'.  These are commands which mainly just alter the access bits on the lock-file of a queue, and they all now use a central routine to do that.  This reduces the amount of code that is run as the priv userid, and eliminates a number of cases where error messages were written while that priv uid was in effect.
As far as users are concerned, there should be no noticable difference
in the new versions.  In case there *is*, the previous implementations
are still there as 'xabort', 'xenable', etc, so they are available for
instant fallback.  If no one reports a problem after a few weeks, then
a later update will remove those x-commands.

Reviewed by:	freebsd-audit and freebsd-print@bostonradio.org
MFC after:	10 days


git-svn-id: file:///home/git/svn/freebsd_mirror/head@98152 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f