Introduce support for Mandatory Access Control and extensible kernel access control.
Label pipe IPC objects, permitting security information to be
maintained at the granularity of the pipe object.  The label is
shared between the two pipe endpoints in the style of the
pipe mutex, and is maintained using similar conventions.  The
label is protected by the pipe mutex.

Obtained from:	TrustedBSD Project
Sponsored by:	DARPA, NAI Labs


git-svn-id: file:///home/git/svn/freebsd_mirror/head@100988 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f