Oops, broke the build there.  Uninline biodone() now that it is non-trivial.
Introduce biowait() function.  Currently there is a race condition and the
mitigation is a timeout/retry.  It is not obvious what kind of locking (if any)
is suitable for BIO_DONE, since the majority of users take are of this
themselves, and only a few places actually rely on the wakeup.

Sponsored by: DARPA & NAI Labs.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@103281 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f