malloc(9) with M_NOWAIT seems to return NULL a lot more than I would have expected under -current.  This is a problem for GEOM because the up/down threads cannot sleep waiting for memory to become free.  The reason they cannot sleep is that paging things out to disk may be the only way we can clear up some RAM.  Nice catch-22 there.
Implement a rudimentary ENOMEM recovery strategy:  If an I/O request
fails with an error code of ENOMEM, schedule it for a retry, and
tell the down-thread to sleep hz/10 to get other parts of the system
a chance to free up some memory, in particular the up-path in GEOM.

All caches should probably start to monitor malloc(9) failures using the new
malloc_last_fail() function, and release when it indicates congestion.

Sponsored by:	DARPA & NAI Labs.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@106338 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f