Define `Sudden_Underflow' when compiling for the Alpha architecture, mainly to avoid getting a SIGFPE signal sent when calling strtod(3) with certain input.
The SIGFPE has been sent because the code was not aware that
a Gradual Underflow is handled in software via traps on the
Alpha architecture, but is not implemented in our Alpha kernel
layer.

With `Sudden_Underflow' defined, strtod(3) should not depend
on Gradual Underflow and adjust its calculations accordingly,
which means that other, more subtle errors than the sending of
SIGFPE could be solved by this.

Discussed with:	bde
PR:		alpha/12623
PR:		alpha/17032
PR:		alpha/43567
MFC after:	7 days


git-svn-id: file:///home/git/svn/freebsd_mirror/head@106918 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f