Reverse the quirk table entry for swizzling on a missing interrupt map; this is now done on all machines except for some known problematic ones.
Add an additional guard to make sure that the interrupt numbers are
in the correct range before swizzling. This should catch any remaining
models for which the swizzle is inappropriate.

Correct the swizzle calculation to account for the fact that the parent
interrupt numbers to be swizzled are 1-based.

Approved by:	re


git-svn-id: file:///home/git/svn/freebsd_mirror/head@107472 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f