Add the `L' option to dump to notify it that it is dumping a live filesystem. To obtain a consistent dump image, dump takes a snapshot of the filesystem and then does a dump of the snapshot. The snapshot is removed when the dump is complete.
Also add an operator warning that the `L' option should be used
if dump is run on a live filesystem without the `L' option being
specified. The alternative would be to silently use a snapshot
any time that a live filesystem is dumped, but this change in
dump semantics seemed too drastic at this time.

Sponsored by:   DARPA & NAI Labs.
Approved by:	re


git-svn-id: file:///home/git/svn/freebsd_mirror/head@107559 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f