Find places to store the previously implicityly passed unit number in the three configuration ioctls which need a unit number.
Add a "ccd.ctl" device for config operations.

Implement ioctls on ccd.ctl which rely on the explicityly passed
unit numbers.

Update ccdconfig to use the new ccd.ctl interface.

Add code to the kernel to detect old ccdconfig binaries, and whine
about it.

Add code to ccdconfig to detect old kernels, and whine about it.

These two compatibility measures will be retained only for a limited
period since they are in the way of GEOM'ification of ccd.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@109421 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f