Complete rewrite of pam_ssh(8).  The previous version was becoming hard to maintain, and had security issues which would have required a major rewrite to address anyway.
This implementation currently starts a separate agent for each session
instead of connecting each new session to the agent started by the first
one.  While this would be a Good Thing (and the old pam_ssh(8) tried to
do it), it's hard to get right.  I'll revisit this issue when I've had a
chance to test some modifications to ssh-agent(1).


git-svn-id: file:///home/git/svn/freebsd_mirror/head@110598 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f