Checkpoint a work in progress: o Use the common pci_* routines in preference to the copied and hacked   routines from an ancient pci.c.
This saves 509 lines in cardbus.c.  More savings to follow when I
convert the resource code over.  In the past when I've done this the
resource code conversion breaks cardbus in subtle ways so I'm doing a
1/2 way checkpoint this time.  cardbus still works for me the same as
it did before.

It also looks like cardbus devices now show up as pci bus devices to
pciconf -l, but maybe that was happening before.

Inspired by a patch from Justin Gibbs many moons ago.  When he
finishes his kobj multiple inheritance work, we can transition the
finished version of this work to that fairly easily.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@110975 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f