- Remove unused cache flushing routines.  These will not necessary work   on future UltraSPARC cpus for which the data cache is not direct mapped. - Move UltraSPARC I and II (spitfire, blackbird, sapphire, sabre) specific   functions to spitfire.c, and add cheetah.c for UltraSPARC III specific   functions.  Initially just cache flushing, but there are a few other   functions that will need to move here. - Add an ipi handler for data cache flushing on UltraSPARC III. - Use function pointers to select the right cache flushing functions based   on cpu_impl.
With this it is possible to boot single user from an mfs root on UltraSPARC
III systems, including spinning up secondary processors.  There is currently
no support for the host to pci bridge, and no documentation for it is
publically available.

Thanks to Oleg Derevenetz for providing access to a system with UltraSPARC
III+ cpus.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@112399 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f