Add module data and version to the atm_subr and reference this info from the (currently) only consumer (en).
Add a sysctl node hw.atm where the atm drivers will hook on their hardware
sysctl sub-trees.

Make atm_ifattach call if_attach and remove the corresponding call to if_attach
from en. Create atm_ifdetach and use that in en.

While the last change actually changes the interface this is not a problem in
practice because the only other consumer of this API is an older LANAI driver
on the net, that is not ready for current anyway.

Reviewed by:	-atm


git-svn-id: file:///home/git/svn/freebsd_mirror/head@114201 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f