Fix rescue build using -j. The problem appears to be make not being able to find the source when the object was specified as <directory>/foo.o.
The fix makes the build go through a make objs before compiling the rest
of the crunchgen. This ensures that the dhclient bits are built in the
correct place where they are picked up for the final compile of rescue.

I'd like to thank dwhite@ and gad@ for helping me track down the problem.

Fast testing box provided by:	phk@ (thanks)


git-svn-id: file:///home/git/svn/freebsd_mirror/head@117956 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f