Don't waste so much space for the latency debugging buffer.  Its size will now need editing except for spot checks.
Changed this buffer from a circular one to a linear one.  This is more
useful for some cases and the sysctl that prints it doesn't support
circular buffers.

Fixed (output) formatting bugs in this sysctl.  An off by 1 error caused
a garbage byte to be returned after annotation of large deltas, and
a race with the writer sometimes caused premature string termination.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@122819 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f