Move all of the recovery thread routines next to each other.
Correct the recovery thread's loop so that it
will terminate properly on shutdown.  We also
clear the recovery_thread proc pointer so that
any additional calls to aic_terminate_recovery_thread()
will not attempt to kill a thread that doesn't
exist.  Lastly, code the loop so that termination
will still be successfull even if the termination
request occurs just prior to us entering the loop
or while the recovery thread is off recovering
commands.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@123679 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f