Use the external clock input for our PLL.
This may not be a generally valid configuration, but neither is relying
on the PCI clock to be stable.

The only currently known and supported hardware is the VPN14x1 from
Soekris, and since it has external clock, we fail safe(r) by using
it.

Unfortunately there is no way to probe this reliably.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@126803 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f