Remove isa compat stuff.
Only cy, bs and wd in the tree still use it.  I have a replacement for
cy that I need to test on ISA and PCI cards.  bs and wd are pc98 only
drivers that appear to no longer be necessary.  I'll be removing them
when I hear back from the pc98 people.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@126996 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f