Remove improper use of if_addrhead in device drivers to check if the link-level address has been initialized already.
The majority of modern drivers never does this and works fine, which
makes me think that the check is totally unnecessary and a residue
of cut&paste from other drivers.

This change is done to simplify locking because now almost none of the
drivers uses this field. The exceptions are "ct" "ctau" and "cx"
where i am not sure if i can remove that part.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@128293 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f