Optimize intr_execute_handlers() by combining the pic_disable_source() and pic_eoi_source() into one call.  This halves the number of spinlock operations and indirect function calls in the normal case of handling a normal (ithread) interrupt.  Optimize the atpic and ioapic drivers to use inlines where appropriate in supporting the intr_execute_handlers() change.
This knocks 900ns, or roughly 1350 cycles, off of the time spent servicing an
interrupt in the common case on my 1.5GHz P4 uniprocessor system.  SMP systems
likely won't see as much of a gain due to the ioapic being more efficient than
the atpic.  I'll investigate porting this to amd64 soon.

Reviewed by:	jhb


git-svn-id: file:///home/git/svn/freebsd_mirror/head@133017 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f