When ISA_PNP_PROBE is called, it will return 0 when it finds a device, ENOENT when there's no PNP ID for this device node, or ENXIO when there is one, but it doesn't match.
In the nonPNP case (as most Alpha systems appear to be), we were
treating the error return as an error, when it should be have ignored
it.  Version 1.9 properly ignored it, but the attach re-write of 1.10
introduced this logic error.

Pointy Hat to: phk (for breaking it then asking me to fix it :-)
Sponsored by: The Voices in Bill Paul's Head, LLC


git-svn-id: file:///home/git/svn/freebsd_mirror/head@134580 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f