Per recent HEADSUP:  Disconnect (old)vinum from the kernel build.
Users should move to the new geom_vinum implementation instead.

The refcount logic which is being added to devices to enable safe module
unloading and the buf/vm work also in progress would require a major rework
of the (old)-vinum code to comply with the new semantics.

The actual source files will not be removed until I have coordinated with
the geomvinum people if they need any bits repo-copied etc.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@135611 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f