When loading GEOM modules, we expect the actual load process to be done by the time that kldload(8) returns.  Satisfy that by making the GEOM module load event -- only when the kernel is !cold -- wait until the GEOM module init function has finished instead of returning immediately.
This is the other half of fixing md(8) (actually, "mfs" in fstab(5))
that is similar to r1.128 of src/sys/dev/md/md.c.  This bug would be
why RAM disks would often fail on boot and the first call to mdconfig(8)
would probably fail.

pjd has ideas for not requiring kldload(8) to work synchronously for
control devices that could make this obsolete.

Silence on:	-arch


git-svn-id: file:///home/git/svn/freebsd_mirror/head@136414 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f