Put the I/O block size in bufobj->bo_bsize.
We keep si_bsize_phys around for now as that is the simplest way to pull
the number out of disk device drivers in devfs_open().  The correct solution
would be to do an ioctl(DIOCGSECTORSIZE), but the point is probably mooth
when filesystems sit on GEOM, so don't bother for now.


git-svn-id: file:///home/git/svn/freebsd_mirror/head@136966 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f