Change gdb_cpu_setreg() to not take the value to which to set the specified register, but a pointer to the in-memory representation of that value. The reason for this is twofold: 1. Not all registers can be represented by a register_t. In particular    FP registers fall in that category. Passing the new register value    by reference instead of by value makes this point moot. 2. When we receive a G or P packet, both are for writing a register,    the packet will have the register value in target-byte order and    in the memory representation (modulo the fact that bytes are sent    as 2 printable hexadecimal numbers of course). We only need to    decode the packet to have a pointer to the register value.
This change fixes the bug of extracting the register value of the P
packet as a hexadecimal number instead of as a bit array. The quick
(and dirty) fix to bswap the register value in gdb_cpu_setreg() as
it has been added on i386 and amd64 can therefore be removed and has
in fact been that.

Tested on: alpha, amd64, i386, ia64, sparc64


git-svn-id: file:///home/git/svn/freebsd_mirror/head@138253 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f