Start the protocol timeouts only after all domains have been initialized completely. For some reason (that I am still curious about) we started to no longer manage to finish the initialization before the timeouts run the first time leading to panics when using uninitialized mutex etc.
The root of this problem is that we currently first link a domain to the
domains list and only later initialize the domain's protocols. This should
be reworked in the future, but with the current API it is not possible in
all situations. We settle with this lazy fix for now.

Tested by:	gnn, ru, myself


git-svn-id: file:///home/git/svn/freebsd_mirror/head@138622 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f