Revert parts of tcp_input.c:1.255 associated with the header predicted cases for tcp_input():
While it is true that the pcbinfo lock provides a pseudo-reference to
inpcbs, both the inpcb and pcbinfo locks are required to free an
un-referenced inpcb.  As such, we can release the pcbinfo lock as
long as the inpcb remains locked with the confidence that it will not
be garbage-collected.  This leads to a less conservative locking
strategy that should reduce contention on the TCP pcbinfo lock.

Discussed with: sam


git-svn-id: file:///home/git/svn/freebsd_mirror/head@139297 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f